apps/example_app/requirements.txt
annotated-types==0.7.0 ; python_full_version == "3.11.9"
anyio==4.4.0 ; python_full_version == "3.11.9"
asgiref==3.8.1 ; python_full_version == "3.11.9"
certifi==2024.7.4 ; python_full_version == "3.11.9"
click==8.1.7 ; python_full_version == "3.11.9"
colorama==0.4.6 ; python_full_version == "3.11.9" and (sys_platform == "win32" or platform_system == "Windows")
deprecated==1.2.14 ; python_full_version == "3.11.9"
dnspython==2.6.1 ; python_full_version == "3.11.9"
email-validator==2.2.0 ; python_full_version == "3.11.9"
fastapi-cli==0.0.4 ; python_full_version == "3.11.9"
fastapi==0.111.1 ; python_full_version == "3.11.9"
h11==0.14.0 ; python_full_version == "3.11.9"
httpcore==1.0.5 ; python_full_version == "3.11.9"
httptools==0.6.1 ; python_full_version == "3.11.9"
httpx==0.27.0 ; python_full_version == "3.11.9"
idna==3.7 ; python_full_version == "3.11.9"
importlib-metadata==7.1.0 ; python_full_version == "3.11.9"
jinja2==3.1.4 ; python_full_version == "3.11.9"
markdown-it-py==3.0.0 ; python_full_version == "3.11.9"
markupsafe==2.1.5 ; python_full_version == "3.11.9"
mdurl==0.1.2 ; python_full_version == "3.11.9"
opentelemetry-api==1.25.0 ; python_full_version == "3.11.9"
opentelemetry-instrumentation-asgi==0.46b0 ; python_full_version == "3.11.9"
opentelemetry-instrumentation-fastapi==0.46b0 ; python_full_version == "3.11.9"
opentelemetry-instrumentation==0.46b0 ; python_full_version == "3.11.9"
opentelemetry-semantic-conventions==0.46b0 ; python_full_version == "3.11.9"
opentelemetry-util-http==0.46b0 ; python_full_version == "3.11.9"
pydantic-core==2.20.1 ; python_full_version == "3.11.9"
pydantic==2.8.2 ; python_full_version == "3.11.9"
pygments==2.18.0 ; python_full_version == "3.11.9"
python-dotenv==1.0.1 ; python_full_version == "3.11.9"
python-multipart==0.0.9 ; python_full_version == "3.11.9"
pyyaml==6.0.1 ; python_full_version == "3.11.9"
rich==13.7.1 ; python_full_version == "3.11.9"
setuptools==70.3.0 ; python_full_version == "3.11.9"
shellingham==1.5.4 ; python_full_version == "3.11.9"
sniffio==1.3.1 ; python_full_version == "3.11.9"
starlette==0.37.2 ; python_full_version == "3.11.9"
typer==0.12.3 ; python_full_version == "3.11.9"
typing-extensions==4.12.2 ; python_full_version == "3.11.9"
uvicorn[standard]==0.30.1 ; python_full_version == "3.11.9"
uvloop==0.19.0 ; (sys_platform != "win32" and sys_platform != "cygwin") and platform_python_implementation != "PyPy" and python_full_version == "3.11.9"
watchfiles==0.22.0 ; python_full_version == "3.11.9"
websockets==12.0 ; python_full_version == "3.11.9"
wrapt==1.16.0 ; python_full_version == "3.11.9"
zipp==3.19.2 ; python_full_version == "3.11.9"
